requirements.txt
# Web Scraping
scrapy>=2.11.0
selenium>=4.15.0
webdriver-manager>=4.0.1

# Data Processing
pandas>=2.1.3
numpy>=1.26.2

# HTTP Requests
requests>=2.31.0

# Data Storage
openpyxl>=3.1.2

# Environment Variables
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.1
